+++ src/measurement_package.h
@@ -4,9 +4,9 @@
 #include "Eigen/Dense"
 
 class MeasurementPackage {
-  
+
 public:
-  long long timestamp_;
+  long timestamp_;
   enum SensorType{LASER,RADAR} sensor_type_;
   Eigen::VectorXd raw_measurements_;
 };
